+++ zebra-network/src/config/tests/vectors.rs
@@ -1,16 +1,9 @@
 //! Fixed test vectors for zebra-network configuration.
 
 use static_assertions::const_assert;
-use zebra_chain::{
-    block::Height,
-    parameters::{
-        subsidy::{
-            FundingStreamReceiver, FundingStreamRecipient, FundingStreams, FUNDING_STREAMS_TESTNET,
-            FUNDING_STREAM_ECC_ADDRESSES_TESTNET, FUNDING_STREAM_MG_ADDRESSES_TESTNET,
-            FUNDING_STREAM_ZF_ADDRESSES_TESTNET, POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_TESTNET,
-        },
-        testnet::{self, ConfiguredFundingStreamRecipient, ConfiguredFundingStreams},
-    },
+use zebra_chain::parameters::{
+    subsidy::FUNDING_STREAMS_TESTNET,
+    testnet::{self, ConfiguredFundingStreams},
 };
 
 use crate::{
@@ -99,42 +92,10 @@ fn default_config_uses_ipv6() {
 fn funding_streams_serialization_roundtrip() {
     let _init_guard = zebra_test::init();
 
-    let fs = vec![
-        ConfiguredFundingStreams::from(&FundingStreams::new(
-            Height(1_028_500)..Height(2_796_000),
-            vec![
-                (
-                    FundingStreamReceiver::Ecc,
-                    FundingStreamRecipient::new(7, FUNDING_STREAM_ECC_ADDRESSES_TESTNET),
-                ),
-                (
-                    FundingStreamReceiver::ZcashFoundation,
-                    FundingStreamRecipient::new(5, FUNDING_STREAM_ZF_ADDRESSES_TESTNET),
-                ),
-                (
-                    FundingStreamReceiver::MajorGrants,
-                    FundingStreamRecipient::new(8, FUNDING_STREAM_MG_ADDRESSES_TESTNET),
-                ),
-            ]
-            .into_iter()
-            .collect(),
-        )),
-        ConfiguredFundingStreams::from(&FundingStreams::new(
-            Height(2_976_000)..Height(2_796_000 + 420_000),
-            vec![
-                (
-                    FundingStreamReceiver::Deferred,
-                    FundingStreamRecipient::new::<[&str; 0], &str>(12, []),
-                ),
-                (
-                    FundingStreamReceiver::MajorGrants,
-                    FundingStreamRecipient::new(8, POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_TESTNET),
-                ),
-            ]
-            .into_iter()
-            .collect(),
-        )),
-    ];
+    let fs = FUNDING_STREAMS_TESTNET
+        .iter()
+        .map(ConfiguredFundingStreams::from)
+        .collect();
 
     let config = Config {
         network: testnet::Parameters::build()
@@ -149,76 +110,4 @@ fn funding_streams_serialization_roundtrip() {
     let deserialized: Config = toml::from_str(&serialized).unwrap();
 
     assert_eq!(config, deserialized);
-}
-
-#[test]
-fn funding_streams_default_values() {
-    let _init_guard = zebra_test::init();
-
-    let fs = vec![
-        ConfiguredFundingStreams {
-            height_range: Some(Height(1_028_500 - 1)..Height(2_796_000 - 1)),
-            // Will read from existing values
-            recipients: None,
-        },
-        ConfiguredFundingStreams {
-            // Will read from existing values
-            height_range: None,
-            recipients: Some(vec![
-                ConfiguredFundingStreamRecipient {
-                    receiver: FundingStreamReceiver::Deferred,
-                    numerator: 1,
-                    addresses: None,
-                },
-                ConfiguredFundingStreamRecipient {
-                    receiver: FundingStreamReceiver::MajorGrants,
-                    numerator: 2,
-                    addresses: Some(
-                        POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_TESTNET
-                            .iter()
-                            .map(|s| s.to_string())
-                            .collect(),
-                    ),
-                },
-            ]),
-        },
-    ];
-
-    let network = testnet::Parameters::build()
-        .with_funding_streams(fs)
-        .to_network()
-        .expect("failed to build configured network");
-
-    // Check if value hasn't changed
-    assert_eq!(
-        network.all_funding_streams()[0].height_range().clone(),
-        Height(1_028_500 - 1)..Height(2_796_000 - 1)
-    );
-    // Check if value was copied from default
-    assert_eq!(
-        network.all_funding_streams()[0]
-            .recipients()
-            .get(&FundingStreamReceiver::ZcashFoundation)
-            .unwrap()
-            .addresses(),
-        FUNDING_STREAMS_TESTNET[0]
-            .recipients()
-            .get(&FundingStreamReceiver::ZcashFoundation)
-            .unwrap()
-            .addresses()
-    );
-    // Check if value was copied from default
-    assert_eq!(
-        network.all_funding_streams()[1].height_range(),
-        FUNDING_STREAMS_TESTNET[1].height_range()
-    );
-    // Check if value hasn't changed
-    assert_eq!(
-        network.all_funding_streams()[1]
-            .recipients()
-            .get(&FundingStreamReceiver::Deferred)
-            .unwrap()
-            .numerator(),
-        1
-    );
 }